refactor: move tests folder to src
